Add support for the magic_load_buffers() function.
The `magic_load_buffers()` function allows for loading pre-compiled Magic
library (one or more) from a location in memory (a buffer), rather than a
file on the file system. This allows for applications which are restricted
to a single directory (e.g., using `chroot`, or other means of sandboxing
like the NaCl or bubblewrap frameworks) or subjected to strict RBAC policies
(enforced through SELinux or AppArmor, etc.) disallowing access outside of
its working directory to load a Magic library and use it.

Signed-off-by: Krzysztof Wilczynski <kw@linux.com>