Build by vendoring libmagic
This commit downloads the `file` library from
http://www.darwinsys.com/file/, builds and installs the files into
`lib/ext`, and links the C extension against it.

At the moment, this isn't staticallly linked because I ran into some
linker issues on macOS, but it could be done later if desired. The
`file` binary is present and could be removed as well.