Fix test failing when MAGIC path set to list of paths
With system libraries, the default MAGIC path may be set to a list of
paths. For example, in Debian and Ubuntu, the path is
`/etc/magic:/usr/share/misc/magic`. This causes libmagic to do further
scans and activates checking of parsed files. We now update the test
to reflect this behavior.