Use mini_portile2 to build libmagic (#5)
* Use mini_portile2 to build libmagic

This will make it easier to use system libraries as Nokogiri does.

* Make static builds work

This commit builds a static library for libmagic and links the
`magic.so` against it. That makes it a little easier to clean up files
that are installed by the library.

* Add --help and clean step

* Add additional configure options

--disable-dependency-tracking: As mentioned in
 https://www.gnu.org/software/automake/manual/html_node/Dependency-Tracking.html,
 is useless for one-time builds.

--disable-silent-rules: As mentioned in
  https://www.gnu.org/software/automake/manual/html_node/Automake-Silent-Rules.html,
  this flag makes it possible to see what `make` is doing with every
  step.

--enable-fsect-man5: Enable file formats in man section 5

* Add Travis CI builds with --disable-static

This will ensure the build works with linking libmagic as a static and
shared library.

* Support building with system library

Using --use-system-libraries with link against the installed
libmagic. magic.h is needed (e.g. via libmagic-dev).

* Update Travis CI to use Ubuntu 18.04 (bionic)

Ubuntu 16.04 (xenial) will reach end-of-life in April 2021, and it has a
more recent version of libmagic (5.32 vs. 5.25).

* Improve cleanup step in build

Explicitly remove the paths we can clean up, leaving the share
directory.

* Fix test failing when MAGIC path set to list of paths

With system libraries, the default MAGIC path may be set to a list of
paths. For example, in Debian and Ubuntu, the path is
`/etc/magic:/usr/share/misc/magic`. This causes libmagic to do further
scans and activates checking of parsed files. We now update the test
to reflect this behavior.

* Add rpath and drop -lmagic substituions

rpath will help enusre we don't link against the static library in
case there isone.

The -lmagic substitutions weren't actually helping because mkmf
have_library was always adding its own -lmagic option.

* Add more files to `rake clobber` list

* Remove unnecessary exports in Travis CI config

All the necessary paths are configured in extconf.rb already.
There is no need to add extra paths.