feat: use the packaged magic.mgc file if it's present
Note that the behavior of Magic#paths has subtly changed here, where we'll only do a lookup the first time (either to magic_getpath_wrapper() or to Magic.default_paths), and from then on the object will use the cached value.

Previously, if the MAGIC env var was set, a lookup would happen every time.